+++ sdk/webpubsub/azure-messaging-webpubsubservice/azure/messaging/webpubsubservice/aio/_patch.py
@@ -78,6 +78,10 @@ class WebPubSubServiceClientConfiguration(Configuration):
         **kwargs: Any
     ) -> None:
         super(WebPubSubServiceClientConfiguration, self).__init__(**kwargs)
+        if kwargs.get("port") and endpoint:
+            if endpoint[-1] == "/":
+                endpoint = endpoint[:len(endpoint) - 1]
+            endpoint = endpoint + ":{}".format(kwargs.pop('port'))
         api_version = kwargs.pop('api_version', "2021-10-01")  # type: str
 
         if endpoint is None: